Compiler fixes.
Mostly related to using a float as an int. I have put explicit
casts where needed and reworked some types to use the proper
target type when declared (float in most cases).

Forgot to declare a few members in class constructors.

+++ engine/src/DrawShapeHelper.cpp
@@ -9,8 +9,8 @@ ATexture* DrawObjectStroke(ARenderer* renderer, FPolygon* target)
     int vertCount = 0;
     vec2** verts = target->GetVertices(&vertCount);
 
-    int minX = verts[0]->x;
-    int minY = verts[0]->y;
+    float minX = verts[0]->x;
+    float minY = verts[0]->y;
 
     // Find the largest X/Y we can subtract to each vertex by finding
     // the smallest x/y coordinate of the vertices.
